[Silver IV] Title: 수열 (Easy), Time: 416 ms, Memory: 64076 KB -BaekjoonHub

+++ 백준/Silver/23827. 수열 （Easy）/수열 （Easy）.java
@@ -8,7 +8,7 @@ public class Main {
     static int N;
     static int[] numbers;
     static long answer;
-    static final int MOD = 1_000_000_007;
+    static final long MOD = 1_000_000_007;
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
@@ -19,8 +19,8 @@ public class Main {
 
         long sum = arrSum(); // 배열 전체 합
         for (int i = 0; i < N; i++) {
-            sum = (sum - numbers[i] + MOD) % MOD; // sum 갱신 시 모듈러 연산 적용
-            answer = (answer + numbers[i] * sum % MOD) % MOD; // 모듈러 연산 후 더하기
+            sum = (sum - numbers[i] + MOD) % MOD; // 음수 방지
+            answer = (answer + numbers[i] * sum % MOD) % MOD;
         }
 
         System.out.println(answer);
@@ -29,7 +29,7 @@ public class Main {
     static long arrSum() {
         long sum = 0;
         for (int i = 0; i < N; i++) {
-            sum = (sum + numbers[i]) % MOD; // 합을 구할 때도 모듈러 연산 적용
+            sum = (sum + numbers[i]) % MOD;
         }
         return sum;
     }
